# perf: avoid closure in BasicDeliver