--- a/src/Channel.ts
+++ b/src/Channel.ts
@@ -140,7 +140,6 @@ export class Channel extends EventEmitter {
     this._conn = conn
     this.id = id
     this.active = true
-    this._invokeHandler = this._invokeHandler.bind(this)
     this._state = {
       emitErrors: emitErrors,
       maxFrameSize: conn._opt.frameMax,
@@ -380,10 +379,19 @@ export class Channel extends EventEmitter {
       }
 
       if (methodFrame.methodId === Cmd.BasicDeliver) {
-        // setImmediate allows basicConsume to resolve first if
-        // basic.consume-ok & basic.deliver are received in the same chunk.
-        // Also this resets the stack trace for handler()
-        setImmediate(this._invokeHandler, uncastMessage as AsyncMessage)
+        const msg = uncastMessage as AsyncMessage
+        const handler = this._state.consumers.get(msg.consumerTag)
+        if (!handler) {
+          // this is a bug; missing handler for consumerTag
+          // TODO should never happen but maybe close the channel here
+        } else {
+          // setImmediate allows basicConsume to resolve first if
+          // basic.consume-ok & basic.deliver are received in the same chunk.
+          // Also this resets the stack trace for handler()
+          // no try-catch; users must handle their own errors
+          setImmediate(handler, msg)
+        }
+
       } else if (methodFrame.methodId === Cmd.BasicReturn) {
         setImmediate(() => {
           this.emit('basic.return', uncastMessage) // ReturnedMessage
@@ -391,19 +399,6 @@ export class Channel extends EventEmitter {
       } else if (methodFrame.methodId === Cmd.BasicGetOK) {
         this._handleRPC(Cmd.BasicGetOK, uncastMessage) // SyncMessage
       }
-    }
-  }
-
-  /** @internal */
-  _invokeHandler(msg: AsyncMessage): void {
-    const handler = this._state.consumers.get(msg.consumerTag)
-    if (!handler) {
-      // this is a bug; missing handler for consumerTag
-      // TODO should never happen but maybe close the channel here
-    } else {
-      // no try-catch; users must handle their own errors
-
-      handler(msg)
     }
   }