@Samuron 2025-08-12
@cody-greene done, please check if I understood you correctly. I've also run a couple of tests and the number of GC pauses stays the same, but all the timings are now different on my machine for whatever reason. I can rerun all the tests if needed, but seems that this version should behave the same in terms of allocations and event better in terms of number of function calls